Made all but 3 views conditionally inclued in MyVideoNav.xml
The only drawbacks are that if the Sets or Recently Added library nodes use one of the conditionally included views it will change to that view for the MovieTitles node as well and the same goes for episodes and TVshows. This can be rectified but not without a slight decrease to available memory. This will be a priority once the way backgrounds are handled has been overhauled.